added destroy and change common package structure
