+++ packages/common/custom-element.ts
@@ -1,0 +1,1 @@
+export * from './src/custom-element/model';
