+++ packages/common/index.ts
@@ -1,6 +1,1 @@
 export * from './src/model/common';
-export * from './src/model/context';
-export * from './src/model/decorator';
-export * from './src/model/elements';
-export * from './src/model/token';
-export * from './src/state/state';
