+++ packages/common/src/destroying/model.ts
@@ -1,0 +1,3 @@
+export interface Cleanable {
+  destroy(): void;
+}
