+++ packages/common/src/destroying/support.ts
@@ -1,0 +1,12 @@
+export class DestroyAction {
+  private closed = false;
+
+  constructor(private callable: VoidFunction) {}
+
+  schedule(): void {
+    if (!this.closed) {
+      requestIdleCallback(this.callable);
+      this.closed = true;
+    }
+  }
+}
