+++ packages/common/src/form-element/model.ts
@@ -1,0 +1,29 @@
+import { HTMLElementCallbacks, HTMLElementConstructor, TargetConstructor } from '../custom-element/model';
+
+export interface FormElementCallbacks extends HTMLElementCallbacks {
+  formAssociatedCallback(form: HTMLFormElement): void;
+  formDisabledCallback(disabled: boolean): void;
+  formResetCallback(): void;
+  formStateRestoreCallback(state: unknown, reason: 'autocomplete' | 'restore'): void;
+}
+
+export interface FormElement
+  extends Pick<
+    ElementInternals,
+    'form' | 'validity' | 'validationMessage' | 'willValidate' | 'checkValidity' | 'reportValidity'
+  > {
+  readonly name: string;
+  readonly type: string;
+}
+
+export type FormTargetCallbacks = Partial<FormElementCallbacks> & object;
+
+export interface FormElementConstructor
+  extends HTMLElementConstructor<HTMLElement & FormElement & FormElementCallbacks> {
+  readonly formAssociated: true;
+}
+
+export type FormElementDecorator = <T extends TargetConstructor<FormTargetCallbacks>>(
+  target: T,
+  context: ClassDecoratorContext,
+) => void;
