+++ packages/common/src/rendering-element/model.ts
@@ -1,24 +1,20 @@
+import { ContextHTMLElement } from '../context/model';
 import {
-  ContextHTMLElement,
   HTMLElementCallbacks,
   HTMLElementConstructor,
   TargetCallbacks,
   TargetConstructor,
-} from '../model/elements';
+} from '../custom-element/model';
 
-export interface RenderingElementCallbacks extends HTMLElementCallbacks {
-}
+export interface RenderingElementCallbacks extends HTMLElementCallbacks {}
 
 export interface RenderingHTMLElement extends ContextHTMLElement, RenderingElementCallbacks {}
-
 
 export interface RenderingTargetCallbacks<R> extends TargetCallbacks {
   render(): R;
 }
 
-export interface RenderingElementConstructor<R> extends HTMLElementConstructor<RenderingTargetCallbacks<R>> {
-  readonly renderAssociated: true;
-}
+export interface RenderingElementConstructor<R> extends HTMLElementConstructor<RenderingTargetCallbacks<R>> {}
 
 export type RenderingElementDecorator<R> = <T extends TargetConstructor<RenderingTargetCallbacks<R>>>(
   target: T,
