+++ packages/common/support.ts
@@ -1,0 +1,2 @@
+export * from './src/destroying/support';
+export * from './src/context/support';
