+++ packages/context/src/use-document.ts
@@ -2,6 +2,5 @@ import { useContext } from './context';
 import { DocumentContext } from './document-context';
 
 export const useDocument = (fn: VoidFunction) => {
-  const context = DocumentContext.create();
-  useContext(context, fn);
+  useContext(DocumentContext.create(), fn);
 };
