+++ packages/custom-element/src/element.ts
@@ -1,4 +1,4 @@
-import { ElementDecorator } from '@diax-js/common';
+import { ElementDecorator } from '@diax-js/common/custom-element';
 import { getElementClass } from './base-element';
 
 export function Element(tagName: string): ElementDecorator {
