+++ packages/form-element/src/form-element.ts
@@ -1,5 +1,5 @@
-import { FormElementDecorator } from "@diax-js/common";
-import { getFormElementClass } from "./base-form-element";
+import { FormElementDecorator } from '@diax-js/common/form-element';
+import { getFormElementClass } from './base-form-element';
 
 export function FormElement(tagName: string): FormElementDecorator {
   return function (target) {
